fix: 修复一处mocel类型错误

+++ src/main/java/com/koala/tools/models/douyin/v1/musicInfo/Music.java
@@ -117,7 +117,7 @@ public class Music implements Serializable {
     @SerializedName("preview_end_time")
     private Integer previewEndTime;
     @SerializedName("preview_start_time")
-    private Integer previewStartTime;
+    private Float previewStartTime;
     @SerializedName("reason_type")
     private Integer reasonType;
     private Boolean redirect;
